git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/Dawn%20of%20Light%20-%20The%20Server@526 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ Teleporters/TeleportNPC.cs
@@ -180,7 +180,7 @@ namespace DOL.GS.GameEvents
 				{
 					LocationExpansion locationex = l as LocationExpansion;
 					bool good = true;
-					if (player.Client.ClientType < (int)locationex.Software)
+					if ((int)player.Client.ClientType < (int)locationex.Software)
 					{
 						SendReply(player, "You don't have the right client type. (" + locationex.Software.ToString() + ")");
 						good = false;
